MIR: add line numbers to templates
The platform does not allow the :lineanchors: option yet, but
this is quite nice already. This way each line gets prepended
by a line number automatically.

These line numbers can be used to say on this link (now
only one template per linke) and say, look at line number X.

Once :lineanchors: works that can even allow deep links in the future.

Nice is also that the copy and paste feature does not include the
numbers, so it is not making the reporter or reviewer life harder.

Fixes #40

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>