Fixes issue #245 add a glossary term for chroot
I hope you find this to be a concise and meaningful definition and if not I look forward to your feedback. This definition has left me with a few questions
Should there be glossary terms for path, root directory and arguments? 
Should chroot be lower case always? I notice that diff is in the glossary in lower case but I'm not sure if that's the house style.
Is there a better place for me to ask/suggest this stuff?